requirements.txt
# Outreach Command Center - Backend Dependencies
# FastAPI Backend for Barton Outreach Core Dashboard

# Core FastAPI framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
psycopg2-binary==2.9.10

# HTTP client for external APIs
requests==2.32.3
httpx>=0.25

# Python data validation
pydantic==2.10.2

# Environment variable management
python-dotenv==1.0.1
